Add simple random attack handler
