+++ src/handlers/message-handler.ts
@@ -2,10 +2,12 @@ import { UUID } from 'node:crypto';
 import { parseRequest } from '../helpers/helpers';
 import { regHandler } from './reg-handler';
 import { createRoomHandler } from './create-room-handler';
-import { IAddShipsData, IAddUserToRoomData, IAttackData, IRegData } from '../types/interface';
+import { IAddUserToRoomData, IRegData } from '../types/interface';
 import { addUserToRoomHandler } from './add-user-to-room-handler';
 import { addShipsHandler } from './add-ships-handler';
 import { attackHandler } from './attack-handler';
+import { IAddShipsData, IAttackData, IRandomAttackData } from '../types/response';
+import { randomAttackHandler } from './random-attack-handler';
 
 export const messageHandler = (message: string, clientId: UUID) => {
   try {
@@ -26,6 +28,9 @@ export const messageHandler = (message: string, clientId: UUID) => {
     }
     if (parsedMessage.type === 'attack') {
       attackHandler(parsedMessage.data as IAttackData, clientId);
+    }
+    if (parsedMessage.type === 'randomAttack') {
+      randomAttackHandler(parsedMessage.data as IRandomAttackData, clientId);
     }
   } catch (err) {
     console.error(err);
