+++ src/types/response.ts
@@ -1,5 +1,6 @@
 import { IShip, IShipPosition, ShipStatus } from './interface';
 import { UserIdType } from '../db/users';
+import { GameIdType } from '../db/app';
 
 export interface IStartGameData {
   ships: IShip[];
@@ -14,4 +15,22 @@ export interface IAttackFeedbackData {
   position: IShipPosition;
   currentPlayer: UserIdType;
   status: ShipStatus;
+}
+
+export interface IAddShipsData {
+  gameId: GameIdType;
+  ships: IShip[];
+  indexPlayer: UserIdType;
+}
+
+export interface IAttackData {
+  gameId: GameIdType;
+  x: number;
+  y: number;
+  indexPlayer: UserIdType;
+}
+
+export interface IRandomAttackData {
+  gameId: GameIdType;
+  indexPlayer: UserIdType;
 }
